pantryai-backend/requirements.txt
Flask>=2.0
python-dotenv>=0.19
supabase>=0.0.1
pillow>=9.0
pytesseract>=0.3
numpy>=1.23
faiss-cpu>=1.7.3
google-generativeai # Changed to get the latest compatible version
gunicorn>=20.0
Flask-Cors>=3.0.10
requests>=2.28.1